requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Werkzeug==3.1.3
SQLAlchemy==2.0.36
yt-dlp
python-dotenv==1.1.0
webvtt-py==0.5.1
# Japanese NLP
sudachipy
sudachidict_core
jamdict
jamdict-data
deep-translator==1.11.4
pykakasi
Flask-WTF
edge-tts
gTTS
pydub
requests
google-generativeai
celery==5.4.0
redis==5.0.1
celery-sqlalchemy-scheduler
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
gunicorn
flask-jwt-extended
